// vhome_mis/src/app/core/api-client-v2/model/serviceUserInputServiceUserInputData.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ServiceUserInputServiceUserInputData { 
    living_space_id: number | null;
    name: string;
    nicknames: Array<string>;
}